fixed an issue where a field took a float but was cast to an int

+++ tModMaker/NEA_solution/GenericBlock.cs
@@ -97,9 +97,9 @@ namespace NEA_solution
 
     internal class create_wings : GenericBlock
     {
-        public int flight_time { get; set; }
-        public int flight_speed { get; set; }
-        public int acceleration { get; set; }
+        public float flight_time { get; set; }
+        public float flight_speed { get; set; }
+        public float acceleration { get; set; }
     }
 
     internal class wing_hover : GenericBlock
